abstracting out sample data into separate file
